Create README.md
This is a web application, which helps to job seekers to find available job vacancies and Recruiters to identify eligible job seekers with the prospect 
of selecting the perfect Employee. The only way to select best-qualified candidate is to have a pool of eligible applicants, which is possible by drawing the interest of individuals in the market.